Consensus working w/ Button
Be as it may, the button doesn't function at all times (problems with Serial connections). Another problem that should be analysed is the following error, it must be included in the report's Experimental Results. Additionally, the Consensus Algorithm is not doing what it should when the external illuminance changes (because there are problems with the reference that comes out of Consensus, the algorithm is doing the right computations, nevertheless, when the illuminance is read from the sensor, rubbish!)